[FINISHED] EnemyFacade
The EnemyFacade is complete!

#EnemyFacadeManager
#EnemyFacadeComponent
#EnemyFacadeUser

These are the scripts required in order to make the Facade work. The manager goes within the GameSystem prefabs. The Component and the User have to go on the prefabs on each enemy.

---MODIFIED FILES---
#EnemyMovement
#ReaperMovement
#GhostMovement

I had to add the following line between each method: enemyFacadeUser?.TriggerFacade("...");
The line in between "" are the action string that the EnemyFacadeManager uses in order to execute the facade. Which calls the SoundSystem, AnimationSystem and VisualEffectSystem that corresponds to said action.